Rename notification provider files to kebab-case
