+++ gui/src/shared/notifications/block-when-disconnected.ts
