+++ gui/src/shared/notifications/notification.ts
@@ -29,13 +29,13 @@ export interface InAppNotificationProvider extends NotificationProvider {
   getInAppNotification(): InAppNotification | undefined;
 }
 
-export * from './accountExpiry';
-export * from './blockWhenDisconnected';
+export * from './account-expiry';
+export * from './block-when-disconnected';
 export * from './connected';
 export * from './connecting';
 export * from './disconnected';
 export * from './error';
-export * from './inconsistentVersion';
+export * from './inconsistent-version';
 export * from './reconnecting';
-export * from './unsupportedVersion';
-export * from './updateAvailable';
+export * from './unsupported-version';
+export * from './update-available';
