+++ gui/src/shared/notifications/update-available.ts
